requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
PyMySQL==1.1.0 